requirements.txt
ruamel.yaml
lxml
